Added custom widget for displaying video items in the queue
A custom widget has been added to QListWidget for displaying YouTube video items in the queue. This allows the video title to be displayed in bold while keeping the URL in a normal font. The custom widget uses a QWidget with a QHBoxLayout to include two QLabel components: one for the video title (styled in bold) and another for the URL (styled normally). The widget height has also been adjusted to ensure readability in line with the font size.

This change improves the UI aesthetics and makes it easier for users to differentiate between video titles and URLs in the queue.